refactor(profiles): split ProfileFormSections into profile-form/* submodules
Extract FieldRow, OptionalSection, ProtonPathField, LauncherMetadataFields,
ProfileSelectorField, TrainerVersionSetField, and helpers into
src/components/profile-form/. Add useProtonDbApply hook under
src/hooks/profile/ that owns ProtonDB state orchestration with 4s status
auto-clear. ProfileFormSections shrinks from 582 to 211 lines; all external
exports preserved.
